Tool validation & raster layer/raster data issue
Updated the ArcGIS tool validation .py to check whether the input raster is projected or not. Issues found when using a raster that was not projected onto 2D space.

Issue where ArcGIS was allowing a user to navigate to a raster data set (as opposed to using a layer from TOC), which was throwing up errors when trying to find the dataSource attribute.